Extract test case expectation of FeedImageDataLoader to a specific protocol that implements XCTestCase
