+++ EssentialApp/EssentialAppTests/FeedImageLoaderCacheDecoratorTests.swift
@@ -38,7 +38,7 @@ final class FeedImageLoaderCacheDecorator: FeedImageDataLoader {
     }
 }
 
-final class FeedImageLoaderCacheDecoratorTests: XCTestCase {
+final class FeedImageLoaderCacheDecoratorTests: XCTestCase, FeedImageTestCase {
     func test_init_doesNotLoadImageData() {
         let (_, loader) = makeSUT()
         
@@ -55,7 +55,6 @@ final class FeedImageLoaderCacheDecoratorTests: XCTestCase {
     }
     
     func test_loadImageData_deliversFeedImageOnLoaderSuccess() {
-        let feedImage = uniqueImage()
         let (sut, loader) = makeSUT()
         
         expect(sut, toCompleteWith: .success(anyData()), when: {
@@ -109,25 +108,6 @@ final class FeedImageLoaderCacheDecoratorTests: XCTestCase {
         trackForMemoryLeaks(loader, file: file, line: line)
         trackForMemoryLeaks(sut, file: file, line: line)
         return (sut, loader)
-    }
-    
-    private func expect(_ sut: FeedImageDataLoader, toCompleteWith expectedResult: FeedImageDataLoader.Result, when action: () -> Void,  file: StaticString = #file, line: UInt = #line) {
-        let exp = expectation(description: "Wait for load expectation")
-        _ = sut.loadImageData(from: anyURL()) { receivedResult in
-            switch (receivedResult, expectedResult) {
-            case let (.success(receivedFeed), .success(expectedFeed)):
-                XCTAssertEqual(receivedFeed, expectedFeed,file: file, line: line)
-            case (.failure, .failure):
-                break
-            default:
-                XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-            }
-            exp.fulfill()
-        }
-        
-        action()
-        
-        wait(for: [exp], timeout: 1.0)
     }
     
     private final class CacheSpy: FeedImageCache {
